feat(image) test added
