+++ src/main/java/org/example/newsfeed/global/config/SecurityConfig.java
@@ -1,7 +1,7 @@
 package org.example.newsfeed.global.config;
+
 import lombok.RequiredArgsConstructor;
-import org.example.newsfeed.global.filter.JwtAuthenticationFilter;
-import org.example.newsfeed.global.util.JwtTokenProvider;
+import org.example.newsfeed.global.filter.JwtFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -15,7 +15,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @RequiredArgsConstructor
 public class SecurityConfig {
 
-    private final JwtTokenProvider jwtProvider;
+    private final JwtFilter jwtFilter;
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -24,7 +24,7 @@ public class SecurityConfig {
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers(
                                 "/api/auth/signup",
-                                "api/auth/login",
+                                "/api/auth/login",
                                 "/signup.html",
                                 "/login.html",
                                 "/mypage.html",
@@ -33,7 +33,7 @@ public class SecurityConfig {
                                 "/favicon.ico").permitAll()
                         .anyRequest().authenticated()
                 )
-                .addFilterBefore(new JwtAuthenticationFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class);
+                .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
     }
