--- src/main/resources/application.properties
@@ -1,17 +1,0 @@
-spring.application.name=newsFeed
-
-
-spring.datasource.url=jdbc:mysql://localhost:3306/newsfeed
-spring.datasource.username=root
-spring.datasource.password=yna51i1E@
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.jpa.hibernate.ddl-auto=create
-
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.use_sql_comments=true
-
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
-
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
