+++ src/main/resources/static/edit-profile.html
@@ -1,0 +1,163 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8" />
+    <title>프로필 수정</title>
+    <style>
+        body {
+            margin: 0;
+            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+            background-color: #f3f4f6;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+        }
+
+        .form-container {
+            background-color: white;
+            padding: 30px 40px;
+            border-radius: 12px;
+            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+            width: 100%;
+            max-width: 500px;
+        }
+
+        h2 {
+            text-align: center;
+            margin-bottom: 24px;
+        }
+
+        label {
+            font-weight: bold;
+            margin-top: 10px;
+            display: block;
+        }
+
+        input, textarea {
+            width: 100%;
+            padding: 10px;
+            margin-top: 6px;
+            margin-bottom: 20px;
+            border: 1px solid #d1d5db;
+            border-radius: 6px;
+        }
+
+        button {
+            width: 100%;
+            background-color: #10b981;
+            color: white;
+            border: none;
+            padding: 12px;
+            font-size: 15px;
+            font-weight: bold;
+            border-radius: 6px;
+            cursor: pointer;
+            transition: background-color 0.3s;
+        }
+
+        button:hover {
+            background-color: #059669;
+        }
+    </style>
+</head>
+<body>
+<div class="form-container">
+    <h2>프로필 수정</h2>
+    <form id="profile-form">
+        <label for="username">이름</label>
+        <input type="text" id="username" name="username" required />
+
+        <label for="bio">소개</label>
+        <textarea id="bio" name="bio" rows="3"></textarea>
+
+        <label for="file">프로필 이미지 업로드</label>
+        <input type="file" id="file" accept="image/*" />
+
+        <label for="profileImage">업로드된 파일명</label>
+        <input type="text" id="profileImage" readonly />
+
+        <button type="submit">수정 완료</button>
+    </form>
+</div>
+
+<script>
+    const baseUrl = "http://localhost:8080";
+    const token = localStorage.getItem("token");
+    const userId = localStorage.getItem("userId");
+
+    if (!token || !userId) {
+        alert("로그인이 필요합니다.");
+        window.location.href = "login.html";
+    }
+
+    window.onload = async () => {
+        const res = await fetch(`${baseUrl}/api/users/${userId}`, {
+            headers: { Authorization: `Bearer ${token}` }
+        });
+
+        const data = await res.json();
+        document.getElementById("username").value = data.username;
+        document.getElementById("bio").value = data.bio || "";
+        document.getElementById("profileImage").value = data.profileImage || "";
+    };
+
+    document.getElementById("profile-form").addEventListener("submit", async (e) => {
+        e.preventDefault();
+
+        const payload = {
+            username: document.getElementById("username").value,
+            bio: document.getElementById("bio").value,
+            profileImage: document.getElementById("profileImage").value
+        };
+
+        const res = await fetch(`${baseUrl}/api/users/profile`, {
+            method: "PUT",
+            headers: {
+                "Content-Type": "application/json",
+                "Authorization": `Bearer ${token}`
+            },
+            body: JSON.stringify(payload)
+        });
+
+        if (res.ok) {
+            alert("✅ 수정 완료!");
+            window.location.href = "mypage.html";
+        } else {
+            alert("❌ 수정 실패: " + res.status);
+        }
+    });
+
+    document.getElementById("file").addEventListener("change", async (e) => {
+        const file = e.target.files[0];
+        if (!file) return;
+
+        const formData = new FormData();
+        formData.append("file", file);
+
+        try {
+            const res = await fetch(`${baseUrl}/api/image/upload`, {
+                method: "POST",
+                headers: {
+                    "Authorization": `Bearer ${token}`
+                },
+                body: formData
+            });
+
+            const text = await res.text();
+            const match = text.match(/파일 업로드 성공: (.+)/);
+            if (match && match[1]) {
+                const uploadedFileName = match[1];
+                document.getElementById("profileImage").value = uploadedFileName;
+                alert("✅ 이미지 업로드 성공!");
+            } else {
+                alert("❌ 업로드 응답을 해석할 수 없습니다.");
+            }
+        } catch (err) {
+            console.error("❌ 업로드 실패", err);
+            alert("❌ 이미지 업로드 실패: " + err.message);
+        }
+    });
+</script>
+</body>
+</html>
