+++ src/main/resources/static/login.html
@@ -1,10 +1,113 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko">
 <head>
-  <meta charset="UTF-8">
-  <title>$Title$</title>
+  <meta charset="UTF-8" />
+  <title>로그인</title>
+  <style>
+    body {
+      font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+      background-color: #f3f4f6;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      height: 100vh;
+      margin: 0;
+    }
+
+    .form-container {
+      background-color: white;
+      padding: 30px 40px;
+      border-radius: 12px;
+      box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+      width: 100%;
+      max-width: 400px;
+    }
+
+    h2 {
+      text-align: center;
+      margin-bottom: 24px;
+    }
+
+    label {
+      font-weight: bold;
+      margin-top: 10px;
+      display: block;
+    }
+
+    input {
+      width: 100%;
+      padding: 10px;
+      margin-top: 6px;
+      margin-bottom: 20px;
+      border: 1px solid #d1d5db;
+      border-radius: 6px;
+    }
+
+    button {
+      width: 100%;
+      background-color: #3b82f6;
+      color: white;
+      border: none;
+      padding: 12px;
+      font-size: 15px;
+      font-weight: bold;
+      border-radius: 6px;
+      cursor: pointer;
+      transition: background-color 0.3s;
+    }
+
+    button:hover {
+      background-color: #2563eb;
+    }
+  </style>
 </head>
 <body>
-$END$
+<div class="form-container">
+  <h2>로그인</h2>
+  <form id="login-form">
+    <label for="email">이메일</label>
+    <input type="email" id="email" required />
+
+    <label for="password">비밀번호</label>
+    <input type="password" id="password" required />
+
+    <button type="submit">로그인</button>
+  </form>
+</div>
+
+<script>
+  const baseUrl = "http://localhost:8080";
+
+  document.getElementById("login-form").addEventListener("submit", async (e) => {
+    e.preventDefault();
+
+    const payload = {
+      email: document.getElementById("email").value,
+      password: document.getElementById("password").value
+    };
+
+    try {
+      const res = await fetch(`${baseUrl}/api/auth/login`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(payload)
+      });
+
+      if (res.ok) {
+        const data = await res.json();
+        localStorage.setItem("token", data.accessToken);
+        localStorage.setItem("userId", data.userId);
+        alert("✅ 로그인 성공!");
+        window.location.href = "mypage.html";
+      } else {
+        const error = await res.text();
+        alert("❌ 로그인 실패: " + error);
+      }
+    } catch (err) {
+      console.error("❌ 에러 발생", err);
+      alert("❌ 요청 실패: " + err.message);
+    }
+  });
+</script>
 </body>
 </html>
