+++ src/main/resources/static/signup.html
@@ -1,0 +1,115 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8" />
+    <title>회원가입</title>
+    <style>
+        body {
+            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+            background-color: #f3f4f6;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+            margin: 0;
+        }
+
+        .form-container {
+            background-color: white;
+            padding: 30px 40px;
+            border-radius: 12px;
+            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+            width: 100%;
+            max-width: 400px;
+        }
+
+        h2 {
+            text-align: center;
+            margin-bottom: 24px;
+        }
+
+        label {
+            font-weight: bold;
+            margin-top: 10px;
+            display: block;
+        }
+
+        input {
+            width: 100%;
+            padding: 10px;
+            margin-top: 6px;
+            margin-bottom: 20px;
+            border: 1px solid #d1d5db;
+            border-radius: 6px;
+        }
+
+        button {
+            width: 100%;
+            background-color: #3b82f6;
+            color: white;
+            border: none;
+            padding: 12px;
+            font-size: 15px;
+            font-weight: bold;
+            border-radius: 6px;
+            cursor: pointer;
+            transition: background-color 0.3s;
+        }
+
+        button:hover {
+            background-color: #2563eb;
+        }
+    </style>
+</head>
+<body>
+<div class="form-container">
+    <h2>회원가입</h2>
+    <form id="signup-form">
+        <label for="email">이메일</label>
+        <input type="email" id="email" required />
+
+        <label for="password">비밀번호</label>
+        <input type="password" id="password" required />
+
+        <label for="username">이름</label>
+        <input type="text" id="username" required />
+
+        <button type="submit">가입하기</button>
+    </form>
+</div>
+
+<script>
+    const baseUrl = "http://localhost:8080";
+
+    document.getElementById("signup-form").addEventListener("submit", async (e) => {
+        e.preventDefault();
+
+        const payload = {
+            email: document.getElementById("email").value,
+            password: document.getElementById("password").value,
+            username: document.getElementById("username").value
+        };
+
+        try {
+            const res = await fetch(`${baseUrl}/api/auth/signup`, {
+                method: "POST",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify(payload)
+            });
+
+            if (res.ok) {
+                alert("✅ 회원가입 성공!");
+                window.location.href = "login.html";
+            } else {
+                const error = await res.text();
+                alert("❌ 회원가입 실패: " + error);
+            }
+        } catch (err) {
+            console.error("❌ 에러 발생", err);
+            alert("❌ 요청 실패: " + err.message);
+        }
+    });
+</script>
+</body>
+</html>
+r
